+++ simulations/job.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 #SBATCH --time=240:00:00
-#SBATCH -J troll_eval
-#SBATCH -o troll_eval.%N.%j.out
-#SBATCH -e troll_eval.%N.%j.err
+#SBATCH -J troll_tee
+#SBATCH -o troll_tee.%N.%j.out
+#SBATCH -e troll_tee.%N.%j.err
 #SBATCH --mem=5G
 #SBATCH --cpus-per-task=1
 #SBATCH --mail-type=ALL
@@ -10,26 +10,15 @@
 
 # Environment
 module purge
-module load bioinfo/Snakemake/7.20.0 # to adapt to your cluster
-module load containers/singularity/3.9.9 # to adapt to your cluster
+module load bioinfo/Snakemake/7.20.0
+module load containers/singularity/3.9.9
 
 # Variables
-CONFIG=config/ressources.yaml
+CONFIG=ressources.yaml
 COMMAND="sbatch --cpus-per-task={cluster.cpus} --time={cluster.time} --mem={cluster.mem} -J {cluster.jobname} -o snake_subjob_log/{cluster.jobname}.%N.%j.out -e snake_subjob_log/{cluster.jobname}.%N.%j.err"
 CORES=200
 mkdir -p snake_subjob_log
 
 # Workflow
-snakemake -s Snakefile --use-singularity -j $CORES --cluster-config $CONFIG --cluster "$COMMAND" --keep-going
+snakemake --use-singularity -j $CORES --cluster-config $CONFIG --cluster "$COMMAND" --keep-going
 
-## Session informations
-echo '########################################'
-echo 'Date:' $(date --iso-8601=seconds)
-echo 'User:' $USER
-echo 'Host:' $HOSTNAME
-echo 'Job Name:' $SLURM_JOB_NAME
-echo 'Job ID:' $SLURM_JOB_ID
-echo 'Number of nodes assigned to job:' $SLURM_JOB_NUM_NODES
-echo 'Nodes assigned to job:' $SLURM_JOB_NODELIST
-echo 'Directory:' $(pwd)
-echo '########################################'